ExceptionHandling 90프로 완성

+++ backend/src/main/java/fullcare/backend/project/service/ProjectService.java
@@ -1,21 +1,19 @@
 package fullcare.backend.project.service;
 
+import fullcare.backend.evaluation.domain.FinalTermEvaluation;
 import fullcare.backend.global.State;
+import fullcare.backend.global.errorcode.ProjectErrorCode;
+import fullcare.backend.global.exceptionhandling.exception.EntityNotFoundException;
 import fullcare.backend.member.domain.Member;
-import fullcare.backend.member.repository.MemberRepository;
 import fullcare.backend.project.domain.Project;
 import fullcare.backend.project.dto.request.ProjectCreateRequest;
 import fullcare.backend.project.dto.request.ProjectUpdateRequest;
 import fullcare.backend.project.dto.response.ProjectListResponse;
-import fullcare.backend.project.dto.response.*;
 import fullcare.backend.project.repository.ProjectRepository;
-import fullcare.backend.projectmember.domain.ProjectMember;
 import fullcare.backend.projectmember.domain.ProjectMemberRole;
 import fullcare.backend.projectmember.domain.ProjectMemberRoleType;
-import fullcare.backend.projectmember.repository.ProjectMemberRepository;
 import fullcare.backend.s3.UploadService;
 import fullcare.backend.util.CustomPageImpl;
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -32,13 +30,10 @@ import java.util.stream.Collectors;
 @Transactional
 public class ProjectService {
     private final ProjectRepository projectRepository;
-    private final MemberRepository memberRepository;
-    private final ProjectMemberRepository projectMemberRepository;
     private final UploadService uploadService;
 
 
-    public Project createProject(Long memberId, ProjectCreateRequest request) {
-        Member member = memberRepository.findById(memberId).orElseThrow();
+    public Project createProject(Member member, ProjectCreateRequest request) {
         Project newProject = Project.createNewProject()
                 .title(request.getTitle())
                 .description(request.getDescription())
@@ -49,26 +44,40 @@ public class ProjectService {
                 .build();
 
         newProject.addMember(member, new ProjectMemberRole(ProjectMemberRoleType.리더, ProjectMemberRoleType.미정));
-
         return projectRepository.save(newProject);
     }
 
     public void updateProject(Long projectId, ProjectUpdateRequest projectUpdateRequest) {
-        Project project = projectRepository.findById(projectId).orElseThrow(() -> new EntityNotFoundException("프로젝트 정보가 없습니다."));
+        Project project = projectRepository.findById(projectId).orElseThrow(() -> new EntityNotFoundException(ProjectErrorCode.PROJECT_NOT_FOUND));
         String imageUrl = project.getImageUrl();
         project.update(projectUpdateRequest);
         uploadService.delete(imageUrl);
     }
 
-
     public Project findProject(Long projectId) {
-        return projectRepository.findJoinPMJoinMemberById(projectId).orElseThrow(() -> new EntityNotFoundException("프로젝트 정보가 없습니다."));
+        return projectRepository.findJoinPMJoinMemberById(projectId).orElseThrow(() -> new EntityNotFoundException(ProjectErrorCode.PROJECT_NOT_FOUND));
     }
 
     public void deleteProject(Long projectId) {
-        Project project = projectRepository.findById(projectId).orElseThrow(() -> new EntityNotFoundException("프로젝트 정보가 없습니다."));
+        Project project = projectRepository.findById(projectId).orElseThrow(() -> new EntityNotFoundException(ProjectErrorCode.PROJECT_NOT_FOUND));
+
+
+        // ? 무슨 용도의 코드인가
+        List<FinalTermEvaluation> finalTermEvaluations = project.getFinalTermEvaluations();
+        for (FinalTermEvaluation fe : finalTermEvaluations) {
+            fe.setProjectNull();
+        }
+
+
         projectRepository.deleteById(projectId);
-        uploadService.delete(project.getImageUrl());
+        if (project.getImageUrl() != null) {
+            uploadService.delete(project.getImageUrl());
+        }
+    }
+
+    public void completeProject(Long projectId) {
+        Project project = projectRepository.findById(projectId).orElseThrow(() -> new EntityNotFoundException(ProjectErrorCode.PROJECT_NOT_FOUND));
+        project.complete();
     }
 
     @Transactional(readOnly = true)
@@ -88,35 +97,4 @@ public class ProjectService {
 
         return new CustomPageImpl<>(content, pageable, pageProject.getTotalElements());
     }
-
-    public List<ProjectMemberListResponse> findProjectMembers(Long projectId) {
-        List<ProjectMember> pmList = projectMemberRepository.findByProjectIdAndProjectMemberRole(projectId, ProjectMemberRoleType.미정);
-
-        List<ProjectMemberListResponse> response = pmList.stream().map(pm -> ProjectMemberListResponse.builder()
-                .id(pm.getMember().getId())
-                .name(pm.getMember().getNickname())
-                .imageUrl(pm.getMember().getImageUrl())
-                .position(pm.getProjectMemberRole().getPosition())
-                .isLeader(pm.isLeader())
-                .build()).collect(Collectors.toList());
-        return response;
-    }
-
-    public List<ProjectMemberListResponse> findApplyList(Long projectId) {
-        List<ProjectMember> pmList = projectMemberRepository.findApplyListByProjectIdAndProjectMemberRole(projectId, ProjectMemberRoleType.미정);
-        List<ProjectMemberListResponse> response = pmList.stream().map(pms -> ProjectMemberListResponse.builder()
-                .id(pms.getMember().getId())
-                .name(pms.getMember().getName())
-                .imageUrl(pms.getMember().getImageUrl())
-                .position(pms.getProjectMemberRole().getPosition())
-                .isLeader(false)
-                .build()).collect(Collectors.toList());
-        return response;
-    }
-
-    public void updateState(Long projectId, State state) {
-        Project project = projectRepository.findById(projectId).orElseThrow();
-        project.updateState(state);
-    }
-
 }
